requirements.txt
ccxt==1.91.75
python-binance==1.0.12
PyMySQL==1.0.2
python-dotenv==0.20.0
PyQt5==5.15.7
plotly==5.9.0
pandas==1.4.2
PyQtChart==5.15.6
pandas==1.4.2